get_faces_from_movie/requirements.txt
dlib~=19.24.1
Pillow~=7.0.0
scipy~=1.10.1
opencv-python~=4.7.0.72
numpy~=1.24.2
matplotlib~=3.7.1